requirements.txt
#Bibliotecas
fastapi
tensorflow
numpy
opencv-python
Pillow
matplotlib
python-multipart
insightface
firebase-admin